refactor: radical refactor for plugin feature
