+++ src/index.ts
@@ -1,7 +1,7 @@
 import { enableMapSet, enablePatches } from 'immer';
 
 export * from './type';
-export * from './class';
+export * from './core';
 export * from './utils/public';
 export * from './const';
