+++ src/type.ts
@@ -13,13 +13,7 @@ export type Patches = Patch[];
 export type PatchesTuple = [Patches, Patches];
 export type PatchesTupleGroup = PatchesTuple[];
 
-export interface HistoryConfig {
-  capacity: number;
-  bufferDebounce: number;
-}
-
-export interface Config {
-  history: HistoryConfig | false;
-  diachrony: boolean;
-  replay: boolean;
+export interface Plugin {
+  name: string;
+  generate(Cls: any): any;
 }
