+++ src/utils/private/index.ts
@@ -1,3 +1,3 @@
 export * from './error';
-export * from './pathCheckCreator';
-export * from './patchesTuple';
+export * from './path-check-creator';
+export * from './patches-tuple';
