+++ src/utils/private/path-check-creator.ts
