+++ src/utils/public/index.ts
@@ -1,6 +1,6 @@
-export * from './assemblePath';
-export * from './bufferDebounceTime';
-export * from './getByPath';
-export * from './isEmptyPath';
-export * from './trimPath';
-export * from './updateDeep';
+export * from './assemble-path';
+export * from './buffer-debounce-time';
+export * from './get-by-path';
+export * from './is-empty-path';
+export * from './trim-path';
+export * from './update-deep';
