+++ src/utils/public/is-empty-path.ts
