+++ src/utils/public/trim-path.ts
